move fetch endpoints to core service
