+++ backend/core/go.mod
@@ -12,6 +12,7 @@ require (
 	common v0.0.0-00010101000000-000000000000
 	firebase.google.com/go/v4 v4.15.2
 	github.com/ably/ably-go v1.2.22
+	github.com/google/uuid v1.6.0
 	github.com/julienschmidt/httprouter v1.3.0
 	github.com/rs/cors v1.10.1
 	github.com/sirupsen/logrus v1.9.3
@@ -47,7 +48,6 @@ require (
 	github.com/golang-jwt/jwt/v4 v4.5.2 // indirect
 	github.com/golang/protobuf v1.5.4 // indirect
 	github.com/google/s2a-go v0.1.8 // indirect
-	github.com/google/uuid v1.6.0 // indirect
 	github.com/googleapis/enterprise-certificate-proxy v0.3.4 // indirect
 	github.com/googleapis/gax-go/v2 v2.14.1 // indirect
 	github.com/joho/godotenv v1.5.1 // indirect
