+++ backend/play/main.go
@@ -29,8 +29,9 @@ func main() {
 	redisPass := env.GetEnvVar("REDIS_PASS")
 	rdb := clients.InitRedisClient(redisHost, redisPort, redisPass)
 
-	// Initialize storage repositories
+	// Initialize redis-based storage repositories
 	repository.InitSessionRepo(implementation.NewSessionRepoRedisImplementation(rdb))
+	repository.InitGameRepo(implementation.NewGameRepoRedisImplementation(rdb))
 
 	// Initialize router
 	router := routing.InitRouter()
