+++ backend/play/routing/handlers/game.go
@@ -3,7 +3,6 @@ package handlers
 import (
 	"net/http"
 	"play/app/session"
-	"sheepshead"
 	"storage/entity"
 	"storage/repository"
 
@@ -24,34 +23,6 @@ func NewGameSession(w http.ResponseWriter, r *http.Request, p httprouter.Params)
 	// Start the session worker.
 	worker.StartWorker()
 	return http.StatusOK, &NewGameSessionResponse{SessionID: entity.ID}
-}
-
-// FetchSessionList returns a list of all active sessions.
-func FetchSessionList(w http.ResponseWriter, r *http.Request, p httprouter.Params) (code int, body any) {
-	sessionRepo := repository.GetSessionRepo()
-	if sessions, err := sessionRepo.GetAll(r.Context()); err != nil {
-		return http.StatusInternalServerError, nil
-	} else {
-		return http.StatusOK, sessions
-	}
-}
-
-// FetchSavedGameList returns any saved games where the host ID matches the provided user ID.
-func FetchSavedGameList(w http.ResponseWriter, r *http.Request, p httprouter.Params) (code int, body any) {
-	uid := p.ByName("uid")
-	gameRepo := repository.GetGameRepo()
-	if games, err := gameRepo.GetAll(r.Context()); err != nil {
-		return http.StatusInternalServerError, nil
-	} else {
-		// filter games by host ID
-		var filteredGames []*sheepshead.Game
-		for _, game := range games {
-			if game.HostID == uid {
-				filteredGames = append(filteredGames, game)
-			}
-		}
-		return http.StatusOK, filteredGames
-	}
 }
 
 type ReviveGameResponse struct {
